promena lozinke prilikom prvog logovanja

+++ MRS2020_T10/src/main/java/main/mrs/service/AdminKCService.java
@@ -18,7 +18,7 @@ public class AdminKCService {
 	@Autowired
 	private AdminKCRepository adminKCRepository;
 	
-	public AdminKC findOne(Long id) {
+	public AdminKC findOne(Integer id) {
 		return adminKCRepository.findById(id).orElseGet(null);
 	}
